ospf6d: create struct ospf6_lsdesc
There are multiple functions that deal with both struct
ospf6_network_lsdesc and struct ospf6_router_lsdesc and resort to using
generic pointers and macros to cast these pointers to the expected type
or access the struct members.

The problem with that pattern is that not only is it more verbose and
harder to read the VERY_LONG_CAPITALIZED_MACRO_NAMES, but they also
break type checking when passed an incorrect pointer, hide endian
conversions and introduce another name for the struct members being
used.

Signed-off-by: Andrew Cooks <acooks.at.bda@gmail.com>