ospf6d: cleanup formatting
This applies selective formatting changes made using tools/indent.py.

ospf6d is inconsistent with the code style in numerous places, which
makes it cumbersome to use indent.py to apply consistent formatting
to any non-trivial change, due to the large number of formatting changes
it makes to otherwisse untouched code.

This commit contains a small subset of the changes indent.py proposes,
and reduces the number of irrelevant changes to consider when staging
a commit.

Signed-off-by: Andrew Cooks <acooks.at.bda@gmail.com>